Rest api

+++ src/main/java/taskmanager/task/api/controller/TaskController.java
@@ -1,7 +1,10 @@
 package taskmanager.task.api.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import taskmanager.task.api.exceptions.TaskNotFoundException;
 import taskmanager.task.api.model.Task;
 import taskmanager.task.api.service.TaskService;
 import java.util.ArrayList;
@@ -9,7 +12,7 @@ import java.util.ArrayList;
 @RestController
 public class TaskController {
 
-    private TaskService taskService;
+    private final TaskService taskService;
 
     @Autowired
     public TaskController (TaskService taskService){
@@ -22,12 +25,32 @@ public class TaskController {
     }
 
     @GetMapping("/task")
-    public Task getTaskByID(@RequestParam int taskID){
+    public ResponseEntity<?> getTaskByID(@RequestParam int taskID) throws TaskNotFoundException {
         Task task = taskService.getTaskByID(taskID);
-        if (task != null){
-            return task;
+        if (task != null) {
+            return ResponseEntity.ok(task);
         }
-        return null;
+        throw new TaskNotFoundException("Task " + taskID + " does not exist!");
+    }
+
+    @GetMapping("/setasdone")
+    public String setTaskAsDone(@RequestParam int taskID) throws TaskNotFoundException {
+        Task task = taskService.getTaskByID(taskID);
+        if (task != null) {
+            return taskService.setTaskAsDone(taskID) ? "Task " + taskID + " successfully marked as done!"
+                    : "Task " + taskID + " is already marked as done!";
+        }
+        throw new TaskNotFoundException("Task " + taskID + " doesn't exist!");
+    }
+
+    @GetMapping("/setasnotdone")
+    public String setTaskAsNotDone(@RequestParam int taskID) throws TaskNotFoundException {
+        Task task = taskService.getTaskByID(taskID);
+        if (task != null) {
+            return taskService.setTaskAsNotDone(taskID) ? "Task " + taskID + " successfully marked as not done!"
+                    : "Task " + taskID + " is already marked as not done!";
+        }
+        throw new TaskNotFoundException("Task " + taskID + " can't be marked as not done since it does not exist!");
     }
 
     @PostMapping("/addtask")
@@ -41,10 +64,14 @@ public class TaskController {
         taskService.modifyTask(taskID, newTask);
         return "Task " + taskID + " modified successfully!";
     }
-
+    
     @DeleteMapping("/delete")
-    public String deleteTask(@RequestParam int taskID){
-        taskService.deleteTask(taskID);
-        return "Task " + taskID + " deleted successfully!";
+    public String deleteTask(@RequestParam int taskID) throws TaskNotFoundException {
+        Task task = taskService.getTaskByID(taskID);
+        if (task != null){
+            taskService.deleteTask(taskID);
+            return "Task " + taskID + " deleted successfully!";
+        }
+        throw new TaskNotFoundException("Task " + taskID + " does not exist, it may have been already deleted or never existed!");
     }
 }
